+++ src/gfx/vao.h
@@ -2,16 +2,17 @@
 #define VAO_H
 
 #include "gfx.h"
+#include "../util/util.h"
 
 struct VAO
 {
-    unsigned int ID;
+    u32 ID;
 };
 
 typedef struct VAO VAO;
 
 VAO vao_init();
 void vao_bind(VAO vao);
-void vao_attr(int index, int count, int size, void* offset);
+void vao_attr(u8 index, u8 count, u32 size, void* offset);
 
 #endif
